refactor(usage): extract pricing primitive into a Node-free leaf module
Review finding: message-cost (imported by the renderer) pulled getModelPricing from
usage-service, whose top-level fs/path/p-limit imports were thus dragged into the web
renderer module graph — a latent web-build risk. Move the pricing table + getModelPricing
into services/model-pricing.ts (just the JSON table, no Node deps); usage-service
re-exports it for back-compat; message-cost imports the leaf directly.
