+++ src/shared/lib/services/model-pricing.ts
@@ -1,0 +1,21 @@
+// Per-million-token pricing for known Claude models.
+// Generated by: npx tsx scripts/fetch-model-pricing.ts
+// When costUSD is present in the JSONL entry, it takes precedence over this table.
+//
+// Leaf module: this carries no Node-only dependencies (just the JSON table) so the
+// renderer can import the pricing primitive (via message-cost) without pulling in
+// usage-service's fs/path/p-limit imports.
+import MODEL_PRICING from './model-pricing.json'
+
+/** Per-model pricing shape (rates are per million tokens). */
+export interface ModelPricing {
+  input: number
+  output: number
+  cacheCreation: number
+  cacheRead: number
+}
+
+/** Returns the per-million-token pricing for a known model, or undefined if unknown. */
+export function getModelPricing(model: string): ModelPricing | undefined {
+  return (MODEL_PRICING as Record<string, ModelPricing>)[model]
+}
