+++ src/shared/lib/services/usage-service.ts
@@ -53,23 +53,11 @@ interface CountedUsage {
   cost: number
 }
 
-// Per-million-token pricing for known Claude models.
-// Generated by: npx tsx scripts/fetch-model-pricing.ts
-// When costUSD is present in the JSONL entry, it takes precedence over this table.
-import MODEL_PRICING from './model-pricing.json'
-
-/** Per-model pricing shape (rates are per million tokens). */
-export interface ModelPricing {
-  input: number
-  output: number
-  cacheCreation: number
-  cacheRead: number
-}
-
-/** Returns the per-million-token pricing for a known model, or undefined if unknown. */
-export function getModelPricing(model: string): ModelPricing | undefined {
-  return (MODEL_PRICING as Record<string, ModelPricing>)[model]
-}
+// Pricing primitive lives in a Node-free leaf module so the renderer (via
+// message-cost) can use it without pulling in this service's fs/path/p-limit deps.
+// Re-exported here so existing usage-service importers keep working.
+export { getModelPricing, type ModelPricing } from './model-pricing'
+import { getModelPricing } from './model-pricing'
 
 /**
  * Calculate cost from token counts using hardcoded pricing.
